App tracks discovery when ServerConnection activity is active.
Turn on discovery whenever the ServerConnection screen is active
so that the discovered name button tracks the current state of
the discovery process. Turn it off when the screen is not active
to save network bandwidth.

While we're at it, greatly improve the readability and clarity
of the ServerConnection screen.