Create the archery-timer-control program
This program is a remote control that uses the network ports of the
archery-timer program to create a remote control panel. The remote
control locates the main timer using ZeroConfg, and allows the user
to connect to the selected timer. At this point, the remote control
can control the timer.